# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This implements the Airport API Reference.",
        "title": "Airport API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/flights": {
            "get": {
                "description": "Retrieves all flight records from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Get all flights",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all flights",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                            }
                        }
                    },
                    "404": {
                        "description": "No flights found",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Add a new flight into the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Add a new flight",
                "parameters": [
                    {
                        "description": "ID of the origin airport",
                        "name": "origin_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "ID of the destination airport",
                        "name": "destination_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Array of flight classes. Each element is a FlightClass object.",
                        "name": "flight_classes",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array"
                        }
                    },
                    {
                        "description": "Flight number",
                        "name": "number",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully added the flight",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {}
                    },
                    "404": {
                        "description": "Origin/Destination airport not found",
                        "schema": {}
                    },
                    "409": {
                        "description": "Flight number conflict",
                        "schema": {}
                    }
                }
            }
        },
        "/flights/{id}": {
            "get": {
                "description": "Retrieves a flight record from the database based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Get a flight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID to get",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved flight",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                        }
                    },
                    "404": {
                        "description": "Flight not found",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Updates a flight record with the provided data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Update a flight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_flights.UpdateFlightBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated flight",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Flight Not Found",
                        "schema": {}
                    },
                    "409": {
                        "description": "Conflict with existing flight number",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Deletes a flight record from the database based on the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Delete a flight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flight ID to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted flight",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                        }
                    }
                }
            }
        },
        "/flights/{number}/{class}": {
            "put": {
                "description": "Updates the price of a flight class with the provided data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flights"
                ],
                "summary": "Update a flight class price",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flight Number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Flight Class",
                        "name": "class",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers_flights.UpdateFlightClassBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated flight class price",
                        "schema": {
                            "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Flight"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Flight or Flight Class Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_krymancer_airport-api-go_pkg_common_models.Flight": {
            "type": "object",
            "properties": {
                "arrival": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "departure": {
                    "type": "string"
                },
                "destination_id": {
                    "type": "integer"
                },
                "flight_class": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.FlightClass"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "origin_id": {
                    "type": "integer"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Ticket"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_krymancer_airport-api-go_pkg_common_models.FlightClass": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "flight_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number_of_seats": {
                    "type": "integer"
                },
                "price_per_seat": {
                    "type": "number"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_krymancer_airport-api-go_pkg_common_models.Lugagge": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "ticket_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_krymancer_airport-api-go_pkg_common_models.Ticket": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "flight": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lugagge": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.Lugagge"
                    }
                },
                "number": {
                    "type": "integer"
                },
                "passenger_id": {
                    "type": "integer"
                },
                "seat_number": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "integer"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "handlers_flights.UpdateFlightBodyRequest": {
            "type": "object",
            "required": [
                "destination_id",
                "flight_classes",
                "number",
                "origin_id"
            ],
            "properties": {
                "destination_id": {
                    "type": "integer",
                    "format": "integer",
                    "example": 2
                },
                "flight_classes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_krymancer_airport-api-go_pkg_common_models.FlightClass"
                    }
                },
                "number": {
                    "type": "string",
                    "format": "string",
                    "example": "FA123"
                },
                "origin_id": {
                    "type": "integer",
                    "format": "integer",
                    "example": 1
                }
            }
        },
        "handlers_flights.UpdateFlightClassBodyRequest": {
            "type": "object",
            "required": [
                "price"
            ],
            "properties": {
                "price": {
                    "type": "number",
                    "format": "float",
                    "example": 100.5
                }
            }
        },
...
    }
}